security: do not export untrusted certs
The prior implementation relied on subcommands of `security(1)` that
all output certificates marked as "untrusted" in keychains.

As there appears to be no native command line application in OSX that
enables differentiation of trusted and untrusted certificates, some
objective c dependencies are now included. The code used requires El
Capitan or newer, and depends on XCode.

The security report was delivered by Eric Hodel. Thank you for your
dilligence, excellent report format, and perfect reporting practice.

The code used to implement the binary introduced by this change was
adapted from code in the Go programming language, and is copyright
"Copyright 2011 The Go Authors" and is covered by the Go LICENSE.

Apologies, but due to lack of time and effort, and also apparent usage,
I also chose not to carry forward the two --skip-* options that
previously allowed users to skip login and/or system certificates.