Chore: Update documentation for ARM based Macs for certs location
Due to the brew prefix being different based on whether the user is
using an Intel based mac or a M1 (ARM) based mac, the initial directory
where the brew packages are is different.

This takes into account to list both possible locations so that if
the user is ever setting the `SSL_CERT_FILE`, they know where to look.